Move and rename tests
