+++ tests/Feature/Actions/SortStreamsByDateActionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Actions;
 
 use App\Actions\SortStreamsByDateAction;
 use App\Models\Stream;
@@ -9,7 +9,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Carbon;
 use Tests\TestCase;
 
-class PrepareStreamsTest extends TestCase
+class SortStreamsByDateActionTest extends TestCase
 {
     use RefreshDatabase;
