+++ tests/Feature/Http/Controllers/CalendarControllerTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Http\Controllers;
 
 use App\Models\Stream;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Carbon;
 use Tests\TestCase;
 
-class CalendarTest extends TestCase
+class CalendarControllerTest extends TestCase
 {
     use RefreshDatabase;
