--- tests/Unit/ExampleTest.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Tests\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-class ExampleTest extends TestCase
-{
-    /** @test */
-    public function exampleTest(): void
-    {
-        $this->assertTrue(true);
-    }
-}
