+++ tests/Unit/FeedTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Unit;
 
 use App\Models\Stream;
 use Carbon\Carbon;
